+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Upstream improvements for Blink engine version detection have been integrated
     - Upstream improvements for browser version detection have been integrated
     - Upstream improvements for Chromium version detection have been integrated
+    - Upstream improvements for "TV-Browser Internet" detection have been integrated
 
 ## v3.10.0 (2024-06-11)